chore(codspeed): run each bench in its own node fork process
