+++ packages/workflow/package.json
@@ -49,6 +49,7 @@
     "@types/luxon": "^3.2.0",
     "@types/md5": "^2.3.5",
     "@types/xml2js": "^0.4.14",
+    "execa": "^5.1.1",
     "tinybench": "^2.8.0"
   },
   "dependencies": {
