+++ packages/workflow/test/benchmark.ts
@@ -1,31 +1,16 @@
-import { Bench } from 'tinybench';
-import { withCodSpeed } from '@codspeed/tinybench-plugin';
+import execa from 'execa';
 import { baseFixtures } from './ExpressionFixtures/base';
-import { evaluate } from './evaluate';
-import type { INodeExecutionData } from '@/index';
-
-function addExpressionEvaluationTasks(bench: Bench) {
-	for (const fixture of baseFixtures) {
-		for (const test of fixture.tests) {
-			if ('error' in test) continue;
-
-			if (test.type === 'evaluation') {
-				const input = test.input.map((d) => ({ json: d })) as INodeExecutionData[];
-				bench.add(fixture.expression, () => evaluate(fixture.expression, input));
-			}
-		}
-	}
-}
 
 async function main() {
-	const bench = withCodSpeed(new Bench());
-
-	addExpressionEvaluationTasks(bench);
-
-	await bench.warmup();
-	await bench.run();
-
-	console.table(bench.table());
+	for (const [fixtureIndex, fixture] of baseFixtures.entries()) {
+		for (const [testIndex, _] of fixture.tests.entries()) {
+			const bench = await execa.node('dist/test/benchTest.js', [
+				fixtureIndex.toString(),
+				testIndex.toString(),
+			]);
+			console.log(bench.stdout);
+		}
+	}
 }
 
 void main();
